app/requirements.txt
streamlit==1.46.1
pandas==2.2.2
python-dotenv==1.0.1
groq==0.37.1
semantic-router==0.1.12
chromadb==1.5.7
sentence-transformers==5.3.0